Extract initial setup code for maintainability
(cherry picked from commit eb1d73f3fbfc493eb56187a7b3457fd697482af3)
